Started Creating Magnetism datapack
This datapack will toggle on and off player magnetism. I want this to be similar to the magnet datapack I made a long time ago, but this one will active for player entities and pull all mob entities as well. Also I'm going to make it accelerate items instead of simply pulling them.